+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Warning for all Users when cookies are currently disabled.
 - Prompt for all Users to use the latest version of Chrome or Firefox.
+- Campers can now be deleted from the admin panel.
 
 ## [1.0.2] - 2019-01-10